remove v1 code
